+++ src/onegov.town/onegov/town/forms/resource.py
@@ -4,6 +4,7 @@ from onegov.form.validators import ValidFormDefinition
 from onegov.town import _
 from onegov.town.utils import mark_images
 from wtforms import StringField, TextAreaField, validators
+from wtforms.fields.html5 import DateField
 from wtforms.widgets import TextArea
 
 
@@ -47,3 +48,27 @@ class ResourceForm(Form):
         self.lead.data = model.meta.get('lead', '')
         self.text.data = model.content.get('text', '')
         self.definition.data = model.definition or ""
+
+
+class ResourceCleanupForm(Form):
+    """ Defines the form to remove multiple allocations. """
+    
+    start = DateField(
+        label=_("Start"),
+        validators=[validators.InputRequired()]
+    )
+
+    end = DateField(
+        label=_("End"),
+        validators=[validators.InputRequired()]
+    )
+
+    def validate(self):
+        result = super(ResourceCleanupForm, self).validate()
+
+        if self.start.data > self.end.data:
+            message = _("The end date must be later than the start date")
+            self.end.errors.append(message)
+            result = False
+
+        return result