+++ src/onegov.town/onegov/town/theme/town_theme.py
@@ -16,7 +16,7 @@ class TownTheme(BaseTheme):
 
     # don't touch this number, it's incremented using bumpversion, so every
     # release will automatically trigger a rebuild of the theme
-    version = '0.4.0'
+    version = '0.5.0'
 
     @property
     def default_options(self):